Redizine and reimplementation for the into tip.
The new requirement could not be fitted in the old implementation. Instead
we use positioning of the tooltip for each sensor and doubled template
for the two modes. Hovering a sensor now shows its tooltip with two buttons.
Positioning is also made somewhat smarter as to avoid being outside the 
screen.